经济系统调整

+++ scripts/citizens/Family.cs
@@ -9,13 +9,13 @@ public class Family
     public List<int> MemberIds = new();
     public int HomeId = -1;
 
-    /// <summary>家庭公产（婚嫁/迁入时注入，日常开销优先扣此）。</summary>
-    public double SharedAssets;
+    /// <summary>家庭公产（文，婚嫁/迁入时注入，日常开销优先扣此）。</summary>
+    public long SharedAssets;
 
-    /// <summary>家庭总资产 = 公产 + 各成员私产。</summary>
-    public double TotalAssets(GameState gs)
+    /// <summary>家庭总资产（文） = 公产 + 各成员私产。</summary>
+    public long TotalAssets(GameState gs)
     {
-        double total = SharedAssets;
+        long total = SharedAssets;
         foreach (var id in MemberIds)
             if (gs.Citizens.TryGetValue(id, out var c))
                 total += c.Money;
